+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [4.21.1](https://github.com/customerio/customerio-android/compare/4.21.0...4.21.1) (2026-09-09)
+
+### Bug Fixes
+
+* **messaginginapp:** preserve spaces in in-app open-url query values ([7fa9996](https://github.com/customerio/customerio-android/commit/7fa999681eda4842e5b8854416985169c9af7585))
+
+## [4.21.0](https://github.com/customerio/customerio-android/compare/4.20.4...4.21.0) (2026-09-09)
+
+### Features
+
+* **inbox:** replace visual inbox empty-state text with a dimmed bell ([#834](https://github.com/customerio/customerio-android/issues/834)) ([a6ff2e2](https://github.com/customerio/customerio-android/commit/a6ff2e2f505201f0c03012274fc057dee10ad463))
+
+## [4.20.4](https://github.com/customerio/customerio-android/compare/4.20.3...4.20.4) (2026-09-09)
+
+### Bug Fixes
+
+* surface the failure reason on in-app error callback ([#828](https://github.com/customerio/customerio-android/issues/828)) ([f9789ae](https://github.com/customerio/customerio-android/commit/f9789aeaff8a1a9c662fc9b587c9dfb0b09cde40))
+
+## [4.20.3](https://github.com/customerio/customerio-android/compare/4.20.2...4.20.3) (2026-09-08)
+
+### Bug Fixes
+
+* **deps:** update Faraday security pin to 1.10.6 ([#845](https://github.com/customerio/customerio-android/issues/845)) ([787d310](https://github.com/customerio/customerio-android/commit/787d3100a5289840be6b6c22e2e60409bc47c06b))
+
 ## [4.20.2](https://github.com/customerio/customerio-android/compare/4.20.1...4.20.2) (2026-08-21)
 
 ### Bug Fixes